move CachingReader classes to their own folder
