+++ src/engine/cachingreader/cachingreader.h
@@ -16,7 +16,7 @@
 #include "track/track.h"
 #include "engine/engineworker.h"
 #include "util/fifo.h"
-#include "engine/cachingreaderworker.h"
+#include "engine/cachingreader/cachingreaderworker.h"
 
 // A Hint is an indication to the CachingReader that a certain section of a
 // SoundSource will be used 'soon' and so it should be brought into memory by
