+++ src/engine/readaheadmanager.cpp
@@ -3,7 +3,7 @@
 
 #include "engine/readaheadmanager.h"
 
-#include "engine/cachingreader.h"
+#include "engine/cachingreader/cachingreader.h"
 #include "engine/controls/loopingcontrol.h"
 #include "engine/controls/ratecontrol.h"
 #include "util/defs.h"
